// product-webapp/webapp/src/app/model/Product.ts
export class Product {
    productName?: string;
    productCategory?: string;
    productSubCategory?: string;
    productBrand?: string;
    productDesc?: string;
    stockQuantity?: number;
    price?: number;
    sellingPrice?: number;
    productImage?: string;
    productId?:  string;
    sellerEmailId?: string;
    review?:string;
  
    // productImage: byte[];
}